HW3: Fixes
