--- bot/src/main/java/edu/java/bot/dto/AddLinkRequest.java
@@ -1,13 +1,0 @@
-package edu.java.bot.dto;
-
-import java.net.URI;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class AddLinkRequest {
-    URI link;
-}
