--- bot/src/main/java/edu/java/bot/dto/LinkUpdate.java
@@ -1,26 +1,0 @@
-package edu.java.bot.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import java.net.URI;
-import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class LinkUpdate {
-    @JsonProperty("id")
-    long id;
-
-    @JsonProperty("url")
-    URI url;
-
-    @JsonProperty("description")
-    String description;
-
-    @JsonProperty("tgChatIds")
-    List<Long> tgChatIds;
-
-}
