--- bot/src/main/java/edu/java/bot/dto/ListLinksResponse.java
@@ -1,17 +1,0 @@
-package edu.java.bot.dto;
-
-import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ListLinksResponse {
-
-    private List<LinkResponse> links;
-
-    private int size;
-
-}
