+++ bot/src/main/java/edu/java/bot/service/command/commands/TrackCommand.java
@@ -2,9 +2,9 @@ package edu.java.bot.service.command.commands;
 
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
-import edu.java.bot.BotUtils;
 import edu.java.bot.dao.LinksDao;
 import edu.java.bot.service.command.Command;
+import edu.java.utils.Utils;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -30,7 +30,7 @@ public class TrackCommand implements Command {
         if (parts.size() < 2) {
             ans = "Please, insert link to track in format \"/track {link}\".";
         } else {
-            if (BotUtils.validateLink(parts.get(1))) {
+            if (Utils.validateLink(parts.get(1))) {
                 var saved =
                     linksDao.saveLink(update.message().chat().id(), update.message().text().split(" ")[1], "");
                 ans = saved ? "Link saved successful." : "Oops! Link was not saved.";
