+++ scrapper/src/main/java/edu/java/clients/GitHubClient.java
@@ -1,8 +1,8 @@
 package edu.java.clients;
 
-import edu.java.Utils;
 import edu.java.dao.LinksDao;
 import edu.java.dto.github.GitHubUpdateDto;
+import edu.java.utils.Utils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
