--- scrapper/src/main/java/edu/java/dto/handlers/AddLinkRequest.java
@@ -1,13 +1,0 @@
-package edu.java.dto.handlers;
-
-import java.net.URI;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class AddLinkRequest {
-    URI link;
-}
