--- scrapper/src/main/java/edu/java/dto/handlers/ApiErrorResponse.java
@@ -1,21 +1,0 @@
-package edu.java.dto.handlers;
-
-import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ApiErrorResponse {
-    String type;
-
-    String code;
-
-    String exceptionName;
-
-    String exceptionMessage;
-
-    List<String> stacktrace;
-}
